refactor: extract shared mock utilities and remove duplicates
- Create mock/utils.ts with common data (titles, avatars, covers, desc, user)
- Export defaultUser, waitTime, fakeList, members as shared utilities
- Refactor mock/user.ts, mock/fakeList.ts to use shared utils
- Refactor account/center/_mock.ts and account/settings/_mock.ts
- Refactor list/basic-list/_mock.ts and list/card-list/_mock.ts
- Remove duplicate mock file requestRecord.mock.js (had redundant APIs)

Removed ~1000 lines of duplicate code across 7 files.
