+++ mock/user.ts
@@ -1,17 +1,5 @@
 import type { Request, Response } from 'express';
-
-const waitTime = (time: number = 100) => {
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve(true);
-    }, time);
-  });
-};
-
-async function getFakeCaptcha(_req: Request, res: Response) {
-  await waitTime(2000);
-  return res.json('captcha-xxx');
-}
+import { waitTime, defaultUser } from './utils';
 
 const { ANT_DESIGN_PRO_ONLY_DO_NOT_USE_IN_YOUR_PRODUCTION } = process.env;
 
@@ -44,55 +32,8 @@ export default {
     res.send({
       success: true,
       data: {
-        name: 'Serati Ma',
-        avatar: 'https://gw.alipayobjects.com/zos/antfincdn/XAosXuNZyF/BiazfanxmamNRoxxVxka.png',
-        userid: '00000001',
-        email: 'antdesign@alipay.com',
-        signature: '海纳百川，有容乃大',
-        title: '交互专家',
-        group: '蚂蚁金服－某某某事业群－某某平台部－某某技术部－UED',
-        tags: [
-          {
-            key: '0',
-            label: '很有想法的',
-          },
-          {
-            key: '1',
-            label: '专注设计',
-          },
-          {
-            key: '2',
-            label: '辣~',
-          },
-          {
-            key: '3',
-            label: '大长腿',
-          },
-          {
-            key: '4',
-            label: '川妹子',
-          },
-          {
-            key: '5',
-            label: '海纳百川',
-          },
-        ],
-        notifyCount: 12,
-        unreadCount: 11,
-        country: 'China',
+        ...defaultUser,
         access: getAccess(),
-        geographic: {
-          province: {
-            label: '浙江省',
-            key: '330000',
-          },
-          city: {
-            label: '杭州市',
-            key: '330100',
-          },
-        },
-        address: '西湖区工专路 77 号',
-        phone: '0752-268888888',
       },
     });
   },
@@ -159,7 +100,7 @@ export default {
     access = '';
     res.send({ data: {}, success: true });
   },
-    'GET /api/500': (_req: Request, res: Response) => {
+  'GET /api/500': (_req: Request, res: Response) => {
     res.status(500).send({
       timestamp: 1513932555104,
       status: 500,
@@ -196,5 +137,8 @@ export default {
     });
   },
 
-  'GET /api/login/captcha': getFakeCaptcha,
+  'GET /api/login/captcha': async (_req: Request, res: Response) => {
+    await waitTime(2000);
+    return res.json('captcha-xxx');
+  },
 };
