+++ src/pages/account/settings/_mock.ts
@@ -1,4 +1,5 @@
 import type { Request, Response } from 'express';
+import { defaultUser } from '../../../../mock/utils';
 
 const city = require('./geographic/city.json');
 const province = require('./geographic/province.json');
@@ -10,70 +11,19 @@ function getProvince(_: Request, res: Response) {
 }
 
 function getCity(req: Request, res: Response) {
-  const province = req.params.province;
+  const provinceKey = req.params.province;
   return res.json({
-    data: city[province as keyof typeof city],
+    data: city[provinceKey as keyof typeof city],
   });
 }
 
 function getCurrentUse(_req: Request, res: Response) {
   return res.json({
-    data: {
-      name: 'Serati Ma',
-      avatar:
-        'https://gw.alipayobjects.com/zos/rmsportal/BiazfanxmamNRoxxVxka.png',
-      userid: '00000001',
-      email: 'antdesign@alipay.com',
-      signature: '海纳百川，有容乃大',
-      title: '交互专家',
-      group: '蚂蚁金服－某某某事业群－某某平台部－某某技术部－UED',
-      tags: [
-        {
-          key: '0',
-          label: '很有想法的',
-        },
-        {
-          key: '1',
-          label: '专注设计',
-        },
-        {
-          key: '2',
-          label: '辣~',
-        },
-        {
-          key: '3',
-          label: '大长腿',
-        },
-        {
-          key: '4',
-          label: '川妹子',
-        },
-        {
-          key: '5',
-          label: '海纳百川',
-        },
-      ],
-      notifyCount: 12,
-      unreadCount: 11,
-      country: 'China',
-      geographic: {
-        province: {
-          label: '浙江省',
-          key: '330000',
-        },
-        city: {
-          label: '杭州市',
-          key: '330100',
-        },
-      },
-      address: '西湖区工专路 77 号',
-      phone: '0752-268888888',
-    },
+    data: defaultUser,
   });
 }
-// 代码中会兼容本地 service mock 以及部署站点的静态数据
+
 export default {
-  // 支持值为 Object 和 Array
   'GET  /api/accountSettingCurrentUser': getCurrentUse,
   'GET  /api/geographic/province': getProvince,
   'GET  /api/geographic/city/:province': getCity,
